+++ ui/src/views/App/Drawer/update.tsx
@@ -32,7 +32,7 @@ const UpdateSLO: React.FC<IProps> = (props) => {
     if (!props.activeSLO) return;
 
     const slo_name = values['slo_name'];
-    const target_slo = parseInt(values['target_slo']);
+    const target_slo = parseFloat(values['target_slo']);
 
     try {
       await _sloService.update(props.activeSLO.slo_name, {
